fixed battle manager

+++ src/modules/BaseEmbed.ts
@@ -5,20 +5,22 @@ import Assets from '../assets';
 const Bundle = Assets.bundle;
 
 export class BaseEmbed extends PagesBuilder {
-  constructor(interaction: CommandInteraction) {
+  constructor(interaction: CommandInteraction, removeable = true) {
     super(interaction);
 
     //remove useless buttons
     this.setPages(new MessageEmbed());
     this.setDefaultButtons([]);
-    this.addComponents(new MessageActionRow().addComponents([
-      new MessageButton().setCustomId('remove_embed').setLabel('Cancel').setStyle('SECONDARY')
-    ])).addTriggers({ 
-      name: 'remove_embed',
-      callback: (inter, componenets) => {
-        setTimeout(()=>interaction.deleteReply().catch(e=>interaction.followUp(Bundle.format(interaction.locale, 'error.delete_message', e))), 1000);
-      }
-    }) 
+    if(removeable) {
+      this.addComponents(new MessageActionRow().addComponents([
+        new MessageButton().setCustomId('remove_embed').setLabel('Cancel').setStyle('SECONDARY')
+      ])).addTriggers({ 
+        name: 'remove_embed',
+        callback: (inter, componenets) => {
+          setTimeout(()=>interaction.deleteReply().catch(e=>interaction.followUp(Bundle.format(interaction.locale, 'error.delete_message', e))), 1000);
+        }
+      })
+    }
   }
